Systematic prototype-drift audit: unify the free, preserve the real
Scripted extern-vs-definition comparison across every function defined
in the repo (~20 substantive drifts after parser noise). Measured each:

Unified (byte-free, MATCH): cd_read_sectors_'s u8*/s32 spelling at
both CD callers, vrealloc's u32 size in SetupSE, AfsFindFile's u32
flags in AfsOpen, leSetEnemy's s16 r in AddEnemy, SetSmokeS's u16
time in ProcItemFire, LoadExecEx's u32 size in exec_process_,
alert-duration knobs.

Retail's own drift, byte-required, annotated in place (nine more, on
top of the two card ones): InitPersistentState and StageSequence
return views in main.c, camera_terrain_pitch_'s short return at both
camera callers, turn_towards_player_'s int view in Think2contact/
Think3firstattack, get_pad_active_ in PauseProc,
update_pressed_buttons' s16 pressed in StateTransition,
tile_sprite_'s s16 in game_over_screen_, set_impact_ex_'s s16 rotate
in DamageControl, and cd_open's stale two-arg prototype in
AfsOpenVolume (the dead mode=0 argument is in the shipped bytes).
GetSpline's pointer-in-s32 SplineRow idiom noted as period style.

The original game plainly shipped with per-TU hand-copied prototypes
and no shared headers; the drifts are part of the bytes.

+++ src/main.exe/AddEnemy.c
@@ -74,7 +74,7 @@ extern u8 str_cancel_2[];               /* cancel */
 extern s32 AdtSelect(char *title, TAdtSelect *menu, s32 mode);
 extern int sprintf(char *buf, char *fmt, ...);
 extern void *memset(void *s, int c, u32 n);
-extern s32 leSetEnemy(s32 type, TThinkType think, s32 x, s32 y, s32 z, s32 r);
+extern s32 leSetEnemy(s32 type, TThinkType think, s32 x, s32 y, s32 z, s16 r);
 
 void AddEnemy(void)
 {
